skills/mysql_connect/requirements.txt
python-dotenv==1.2.2
pymysql==1.1.2
